# docs/aidlc/05-api-specification/Fantasy EPL League Manager — OpenAPI Specification v1.0.yaml
openapi: 3.0.3
info:
  title: Fantasy EPL League Manager API
  version: "1.0"
  description: |
    # Fantasy EPL League Manager
    ## OpenAPI Specification — Version 1.0

    **Document Status:** Baseline API Specification

    **Inputs:**
    - Architecture and Domain Model v1.15 (`Fantasy EPL League Manager — Architecture and Domain Model v1.15.md`) — Section 9's endpoint sketches (§9.2), the aggregate/entity model (§6), logical data architecture (§8), and cross-cutting API conventions (§9.3) that this document turns into concrete request/response DTOs and validation rules.
    - Business Requirements Document v1.17 (`Fantasy EPL League Manager — Business Requirements Document v1.17.md`) — authoritative business rules, cited throughout by `BR-###`.
    - Epic and Feature Backlog v1.10 (`Fantasy EPL League Manager — Epic and Feature Backlog v1.10.md`) — feature IDs (`F-###.#`) cited on each operation group to show which feature it serves and in what build-sequence position (Backlog §5).

    **Purpose:** Turn Architecture §9's illustrative endpoint list into a complete, implementation-ready OpenAPI contract — every resource, request/response schema, validation rule, and error shape needed to start writing controllers and client code. This is the artifact Architecture §16 lists as recommended-next-artifact #3.

...

...

    ## Conventions (Architecture §9.1/§9.3)

    - REST over HTTPS only; JSON request/response bodies.
...

    ## Version History

    ### Version 1.0

...
servers:
...
security:
  - bearerAuth: []
...
paths:
...
  /users/me:
    get:
      tags: [Identity]
      operationId: getCurrentUser
      summary: Get the caller's own User/UserProfile (BR-005, BR-015 — private fields included, since this is self-access)
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserSelfDto'
    put:
      tags: [Identity]
      operationId: updateUsername
      summary: Change the caller's username (F-001.3, BR-003, BR-004, BR-265, BR-266, BR-270)
      description: |
        Opens a new `UsernameHistory` row and closes the previous one (BR-326) so historical screens keep resolving
        the display name that was active at the time of each historical record.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username]
              properties:
                username:
                  $ref: '#/components/schemas/Username'
      responses:
        '200':
          description: Username changed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserSelfDto'
        '400':
          $ref: '#/components/responses/ValidationError'
        '409':
          description: Username already taken by another active User (BR-004).
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
...
  /admin/users/{userId}/reactivate:
    post:
      tags: [Identity]
      operationId: reactivateUser
      summary: System Administrator reactivates a retired User (BR-300, BR-301)
      description: |
        System-Administrator-only (ADR-007) — not reachable through any League-facing screen or self-service flow.
        If the User's prior username has since been taken by another active User, `newUsername` is required in the same request.
      x-authorization: Caller must have `User.IsSystemAdministrator = true`.
      parameters:
        - $ref: '#/components/parameters/userId'
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                newUsername:
                  $ref: '#/components/schemas/Username'
      responses:
        '200':
          description: User reactivated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserSelfDto'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          description: Prior username is taken and newUsername was not supplied.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'

  # ────────────────────────────────────── League & Season ──────────────────────────────────────
...
  /drafts/{draftId}/picks:
    post:
      tags: [Draft]
      operationId: makeDraftPick
      summary: Submit the current turn's player selection (F-005.2, BR-052, BR-059, BR-191, BR-192, AP-009, AP-010)
      description: |
        Atomic: the `DraftSelection` insert and the `SquadPlayer` creation happen in one transaction guarded by the
        partial-unique-ownership index, so two concurrent picks for the same player cannot both succeed — the loser
        receives `409` with `errorCode: player_already_owned`, not a raw constraint violation.
      x-authorization: Caller must own the FantasyTeam whose turn it currently is.
      parameters:
        - $ref: '#/components/parameters/draftId'
        - $ref: '#/components/parameters/idempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [playerId]
              properties:
                playerId: { type: string, format: uuid }
      responses:
        '201':
          description: Pick recorded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DraftSelection'
        '409':
          description: Player already owned by another FantasyTeam in this League/Season (BR-035, BR-191).
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
...
  /leagues/{leagueId}/seasons/{seasonId}/fantasy-teams/{fantasyTeamId}/replacement-opportunities:
    parameters:
      - $ref: '#/components/parameters/leagueId'
      - $ref: '#/components/parameters/seasonId'
      - $ref: '#/components/parameters/fantasyTeamId'
    get:
      tags: [Draft]
      operationId: listReplacementOpportunities
      summary: List a FantasyTeam's unspent replacement-selection opportunities (F-006.4, BR-063–BR-065, BR-287, BR-307)
      description: An opportunity never expires on its own (BR-307) — it stays listed until spent via a Replacement-type draft pick.
      x-authorization: Caller must own fantasyTeamId, or be the League Administrator.
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ReplacementOpportunity'
...
  /leagues/{leagueId}/seasons/{seasonId}/fantasy-teams/{fantasyTeamId}/season-goal-prediction:
    parameters:
      - $ref: '#/components/parameters/leagueId'
      - $ref: '#/components/parameters/seasonId'
      - $ref: '#/components/parameters/fantasyTeamId'
    get:
      tags: [Competition]
      operationId: getSeasonGoalPrediction
      summary: Get a FantasyTeam's Season Goal Prediction, if submitted (F-010.3, BR-126–BR-135)
      x-authorization: Caller must own fantasyTeamId, or be an active member of its League.
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SeasonGoalPrediction'
        '404':
          description: Not yet submitted.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
    put:
      tags: [Competition]
      operationId: submitSeasonGoalPrediction
      summary: Submit or update the Season Goal Prediction before it locks (F-010.3, BR-126–BR-128, BR-299)
      description: Locks at Season start under normal submission (BR-127/BR-128), or at time of submission under the late-submission fallback (BR-299).
      x-authorization: Caller must own fantasyTeamId.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [predictedEplGoals]
              properties:
                predictedEplGoals: { type: integer, minimum: 0 }
      responses:
        '200':
          description: Recorded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SeasonGoalPrediction'
        '409':
          description: Prediction is already locked (BR-127).
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'

  # ─────────────────────────────────────── Corrections & Administration ───────────────────────────────────────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    leagueId:
      name: leagueId
      in: path
      required: true
      schema: { type: string, format: uuid }
    seasonId:
      name: seasonId
      in: path
      required: true
      schema: { type: string, format: uuid }
    fantasyTeamId:
      name: fantasyTeamId
      in: path
      required: true
      schema: { type: string, format: uuid }
...
    draftId:
      name: draftId
      in: path
      required: true
      schema: { type: string, format: uuid }
...
    userId:
      name: userId
      in: path
      required: true
      schema: { type: string, format: uuid }
...
    idempotencyKey:
      name: Idempotency-Key
      in: header
      description: Client-generated UUID so a retried request cannot double-submit (BR-237).
      schema: { type: string, format: uuid }
...
  responses:
    ValidationError:
      description: Request failed shape/format validation (BR-166).
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
...
    Forbidden:
      description: Authenticated, but the caller fails the operation's object-level authorization check (BR-163, BR-254).
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
...
  schemas:
    ProblemDetails:
      type: object
      description: RFC 7807-style error body (Architecture §9.3), extended with a stable machine-readable errorCode.
      properties:
        type: { type: string, format: uri }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        instance: { type: string }
        errorCode:
          type: string
          example: player_already_owned
        correlationId: { type: string, description: Propagated into logs (BR-241). }
    Username:
      type: string
      minLength: 3
      maxLength: 24
      pattern: '^[A-Za-z0-9_]+$'
      description: Unique among active Users (BR-004); reused screens display it verbatim (BR-003). Length/charset are an implementation default, not yet a cited BR — confirm before locking a physical CHECK constraint.
...
    UserStatus:
      type: string
      enum: [Active, Retired]
    UserSelfDto:
      type: object
      description: Self-access DTO — includes private fields (BR-015 restricts league-facing APIs only, not the caller's own /users/me).
      properties:
        userId: { type: string, format: uuid }
        username:
          $ref: '#/components/schemas/Username'
        email: { type: string, format: email }
        status:
          $ref: '#/components/schemas/UserStatus'
        defaultIconId: { type: string, format: uuid }
        isSystemAdministrator: { type: boolean, description: "Never settable via any self-service API (BR-300)." }
        createdAt: { type: string, format: date-time }
...
    DraftSelection:
      type: object
      properties:
        draftSelectionId: { type: string, format: uuid }
        draftId: { type: string, format: uuid }
        fantasyTeamId: { type: string, format: uuid }
        playerId: { type: string, format: uuid }
        round: { type: integer }
        pickNumber: { type: integer }
        selectedAt: { type: string, format: date-time }
        isMakeupPick: { type: boolean, description: "BR-282" }
...
    ReplacementOpportunity:
      type: object
      properties:
        opportunityId: { type: string, format: uuid }
        fantasyTeamId: { type: string, format: uuid }
        sourcePlayerId: { type: string, format: uuid, description: "The player whose exit/injury generated this opportunity." }
        grantedAt: { type: string, format: date-time }
        grantReason:
          type: string
          enum: [EplExit, SeasonEndingInjury]
        spentAt: { type: string, format: date-time, nullable: true }
...
    SeasonGoalPrediction:
      type: object
      properties:
        predictionId: { type: string, format: uuid }
        seasonId: { type: string, format: uuid }
        fantasyTeamId: { type: string, format: uuid }
        predictedEplGoals: { type: integer }
        submittedAt: { type: string, format: date-time }
        lockedAt: { type: string, format: date-time }
        finalActualGoals: { type: integer, nullable: true }
        finalAbsoluteDifference: { type: integer, nullable: true, description: "BR-131/BR-135, computed once at season end." }
...